+++ hmeg/vocabs/nanolex.toml
@@ -1,0 +1,93 @@
+name = "Minilex"
+verbs = [
+    "arrive", "ask",
+    "become", "begin", "believe", "borrow", "bring", "buy",
+    "clean", "come", "continue", "cry",
+    "damage", "do", "drop",
+    "eat",
+    "fall", "feel", "find", "finish", "forget", "free",
+    "have", "head", "hold", "hope", "hurt",
+    "joke",
+    "keep", "know",
+    "laugh", "learn", "leave", "lend", "let", "lie", "like", "listen", "live", "look", "look for", "love",
+    "make", "mean", "meet",
+    "name", "need",
+    "pay", "place", "play", "put",
+    "read", "remember",
+    "say", "see", "sell", "send", "show", "shut", "sing", "speak", "stand", "stay", "stop", "suggest",
+    "take", "talk", "teach", "think", "travel",
+    "understand", "use",
+    "wait", "walk", "want", "work", "worry", "write"
+]
+
+nouns = [
+    "accident", "address", "air", "anybody", "anything", "apartment",
+    "bag", "beginning", "book", "bottom", "boy", "bread", "brother", "buy",
+    "change", "child", "clothes", "cold", "colour", "country", "cry",
+    "damage", "daughter", "day",
+    "earth", "evening", "everybody", "everything", "eye",
+    "fall", "family", "father", "flat", "food", "foot", "friend", "front", "fun",
+    "girl", "ground",
+    "half", "hand", "head", "help", "home", "hope", "hour", "house", "husband",
+    "joke", "journey",
+    "key",
+    "lady", "language", "laugh", "left", "lie", "life", "light", "love",
+    "man", "mistake", "moment", "money", "month", "morning", "mother", "mouth",
+    "name", "need", "news", "newspaper", "night", "nobody", "nothing", "now",
+    "other",
+    "page", "pain", "paper", "part", "pen", "pencil", "people", "person", "picture", "place",
+    "question",
+    "reason", "relative", "right", "road", "room",
+    "school", "shop", "show", "side", "sister", "sleep", "somebody", "something", "son", "sound", "stand", "stop", "street", "suitcase", "sun", "surprise",
+    "talk", "tax", "teacher", "thing", "ticket", "time", "today", "tomorrow", "top", "town", "travel", "true",
+    "use",
+    "voice",
+    "walk", "watch", "water", "way", "weather", "week", "wife", "will", "woman", "word", "work", "world", "worry",
+    "year", "yesterday"
+]
+
+adjectives = [
+    "angry", "another", "any",
+    "bad", "beautiful", "boring",
+    "calm", "certain", "cheap", "cheerful", "clean", "clear", "cold", "common",
+    "dangerous", "dead", "dear", "different", "difficult", "dirty", "dry",
+    "each", "easy", "every", "expensive",
+    "famous", "fast", "finished", "first", "free", "full", "fun", "funny",
+    "glad", "good", "great",
+    "happy", "heavy", "high", "hot",
+    "ill", "important", "impossible", "interesting",
+    "kind",
+    "last", "late", "left", "light", "little", "low",
+    "mean",
+    "near", "necessary", "new", "next", "nice", "no",
+    "open", "other",
+    "possible", "pretty",
+    "quiet", "relative", "right",
+    "safe", "same", "several", "short", "shut", "slow", "small", "some", "sorry", "strange", "strong", "stupid", "such", "surprised",
+    "tall", "terrible", "tired",
+    "urgent", "usual",
+    "warm", "whole", "wonderful", "worst", "wrong",
+    "young"
+]
+
+adverbs = [
+    "angrily",
+    "badly", "beautifully", "boringly",
+    "calmly", "certainly", "cheaply", "cheerfully", "cleanly", "clearly", "coldly", "commonly",
+    "dangerously", "deadly", "dearly", "differently", "difficultly", "dirtily", "dryly",
+    "easily", "every", "expensively",
+    "famously", "fastly", "firstly", "freely", "fully", "funnily",
+    "gladly", "greatly",
+    "happily", "heavily", "highly", "hotly",
+    "importantly", "impossibly", "interestingly",
+    "kindly",
+    "lastly", "lately", "lightly",
+    "nearly", "necessarily", "newly", "next", "nicely",
+    "openly",
+    "possibly", "prettily",
+    "quietly", "relatively", "rightly",
+    "safely", "shortly", "slowly", "sorrily", "strangely", "strongly", "stupidly", "surprisingly",
+    "terribly", "tiredly",
+    "urgently", "usually",
+    "warmly", "well", "wonderfully", "wrongly"
+]
