+++ tests/test_usecases.py
@@ -1,7 +1,7 @@
 import random
 import unittest
 
-from hmeg import utils
+from hmeg import usecases as uc
 from hmeg.entities import VocabularyPlaceholders
 from hmeg.vocabulary import Vocabulary
 
@@ -16,20 +16,20 @@ class TestUtils(unittest.TestCase):
         random.seed(42)
 
         with self.subTest("Empty input"):
-            self.assertEqual(utils.apply_vocabulary("", self.vocab), "")
+            self.assertEqual(uc.apply_vocabulary("", self.vocab), "")
 
         with self.subTest("No placeholders"):
             s = "abcde123^%&#$}{"
-            self.assertEqual(utils.apply_vocabulary(s, self.vocab), s)
+            self.assertEqual(uc.apply_vocabulary(s, self.vocab), s)
 
         with self.subTest("Multiple placeholders"):
             s = "[{verb}][{verb}][{verb}]"
-            res = utils.apply_vocabulary(s, self.vocab)
+            res = uc.apply_vocabulary(s, self.vocab)
             with self.assertRaises(ValueError):
                 res.index(res[:3], 1)  # no repetitions in the generated string
 
         with self.subTest("All placeholders"):
             all_placeholders = " ".join(VocabularyPlaceholders.to_list())
-            res = utils.apply_vocabulary(all_placeholders, self.vocab)
+            res = uc.apply_vocabulary(all_placeholders, self.vocab)
             for placeholder in VocabularyPlaceholders.to_list():
                 self.assertTrue(placeholder not in res)
